feat(capacitor): rename and update `appId`
from Ionic/Capacitor boilerpalte
	to `com.zolfinak.app`
